+++ evalai/login.py
@@ -8,13 +8,13 @@ from evalai.utils.config import AUTH_TOKEN_PATH, AUTH_TOKEN_DIR
 
 
 @click.group(invoke_without_command=True)
-@click.argument("--username", prompt=True)
-@click.argument("--passowrd", prompt=True)
 @click.pass_context
 def login(ctx, username, password):
     """
     Login to EvalAI and save token.
     """
+    username = click.prompt("username", type=str, hide_input=False)
+    password = click.prompt("Enter password", type=str, hide_input=True)
     token = get_user_auth_token_by_login(username, password)
 
     if os.path.exists(AUTH_TOKEN_PATH):
